feat: Use jumps rather than ifs in MIR
Because `match` statements compile to too complex MIR code without jumps
or another rich syntax. E.g. In F#,

```fs
    match t with
    | Some (x, _) when p(x) -> x
    | _ -> 0
```

should compile to a kind of instructions like this:

    1. check if t is Some,
    2. bind t.0 to x, and
    3. check if p(x) is true

noting that we cannot access to `t.0` before checking t is some
or it should cause memory access violation,
and we want to bind `x` before evaluating `p(x)` for brevity.