fix: Fix an inference bug (illegal generalization)
Inner functions must NOT be generic on meta types
that are defined outside the inner function.

We can check if a meta type is defined inside the function or outside
by comparing a type serial with the last type serial of the time
when the inference enters the inner function declaration.

See `let_fun_generic_nesting` test case for minimal reproduction.