fix: Resolve transitive references correctly in ClosureConversion
See `let_fun_closure.fs` for the reason.
We did need the *closure* of the reference graph.

This solution maybe adhoc: mutually recursion is a root cause of bugs.
We should think twice about it to handle in more sophisticated way.
It seems closure conversion should use the result of name resolution.