refactor: Use local functions in arms to reduce stack frame size
Stack overflow occurs when repeatedly call functions which consume
the call stack with large stack frame.

To reduce the size of stack frame of the recursive functions
(astToHir etc.), add rewrote match arms to just call a local funtion
and work in the local function.

This conversion should be done by the compiler itself. I don't implement
the feature just now because self-hosting is currently prior to others.