feat: Reduce use of stack to sort
Modify the merge function to be tail-call recursive. Without this,
it consumes the stack by O(N) size to sort a long (N-length) list,
which runs into a stack overflow.

Note the go function doesn't need to be tail-call because the sort
function calls it at most O(log N) times.