feat: Initialize local variable types on first occurrence in a pattern
To reduce the number of meta types.

Note inferIrrefutablePat now infers that pattern even if it's refutable.
This ensures all variables are initialized before use
and prevents assertion violations in inferVarExpr.
Generally, type checker should process every piece of patterns
and expressions to ensure this kind of invariants, so that's modified.